// src/types/City.ts
import {Location} from './location';

export type City = {
  name: string;
  location: Location;
};

export type CityList = {
  Paris: string;
  Cologne: string;
  Brussels: string;
  Amsterdam: string;
  Hamburg: string;
  Dusseldorf: string;
};

export type CityName = 'Paris' | 'Cologne' | 'Brussels' | 'Amsterdam' | 'Hamburg' | 'Dusseldorf';